University of Michigan Panel Study of Income Dynamics (PSID)

Sample of 753 between the ages of 30 and 60 in 1975 (PSID interview year 1976)

Usage

mroz

Format

A data frame with 753 observations on 22 variables.

inlf

Binary: in labor force in 1975

hours

Hours worked in 1975

kidslt6

Number of children less than 6 years old

kidsge6

Number of children 6 years old or older

age

Woman's age in years

educ

Woman's years of education

wage

Woman's estimated average hourly earnings

repwage

Representative wage in 1976 (interview year)

hushrs

Hours worked by husband in 1975

husage

Husband's age in years

huseduc

Husband's years of education

huswage

Husband's estimated average hourly earnings

faminc

Total family income

mtr

Woman's Federal marginal income tax rate

motheduc

Mother's years of education

fatheduc

Father's years of education

unem

Unemployment rate in county of residence

city

Binary: lives in Standard Metropolitan Statistical Area (SMSA)

exper

Years of labor market experience

nwifeinc

Income from other sources than the wife's labor, in thousands

lwage

log(wage)

expersq

exper^2

Source

https://cran.r-project.org/package=wooldridge

References

Wooldridge, J. M. (2020). Introductory Econometrics: A Modern Approach. 7th Edition. Boston, MA: Cengage Learning.

Mroz, T. A. (1987). "The Sensitivity of an Empirical Model of Married Women's Hours of Work to Economic and Statistical Assumptions." Econometrica 55, 765-799.