401(k) Participation Rates

A dataset containing participation rates and firm characteristics used in Papke and Wooldridge (1996).

pw401k

Format

A data frame with 4734 rows and 5 variables:

prate

participation rate, proportion

mrate

matching rate, proportion

totemp

total number of employees

age

years the plan has been in place

sole

binary indicating if it's sole plan offered by employer

Source

Papke, L. E. and Wooldridge, J. M. (1996).

References

Papke, L. E., & Wooldridge, J. M. (1996). "Econometric methods for fractional response variables with an application to 401(k) plan participation rates." Journal of Applied Econometrics, 11(6), 619-632. doi:10.1002/(SICI)1099-1255(199611)11:6<619::AID-JAE418>3.0.CO;2-1